Set foundation with NamedTuple to(see description)
Set foundation with NamedTuple to save any data at start of program
(which is pulled from the api) for later in program. Idea e.g.: Set
background color of each item with help of the rarity grades. Have to
have which rarity the item has in the middle of the program.