Now with video support
The program will now search for videos for items, which occurre for the
first time in the fortnite shop. They will be sent over the telegram
api (like the message and picture). The videos are treated as a seperate
part to the picture. Each time the program runs it will check if it
already has stored a item video in the previouse runs. If so, it won't
send the video again. If not so, it'll store the video and send it. Each
shift in day it will forget the previouse videos (but
nevertheless shouldn't send it beacause the occurrence changed). The
program !doesn't compare if the video of a item also changes throught
the day (like with the picture), just if it exists or not. To make this
happen the id provided for each video by the fortnite api is used for each video.

Other improvments/changes:
  - updated names of files
  - split up code in multiple files for a better overview
  - added settings